refactor: MainActivity mvp 패턴 적용
