+++ app/src/main/java/woowacourse/movie/conrtract/MainContract.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.conrtract
+
+interface MainContract {
+    interface View {
+        fun setUpMovieContentListAdapter()
+    }
+
+    interface Presenter {
+        fun saveMovieContent()
+    }
+}
