+++ app/src/main/java/woowacourse/movie/view/BaseActivity.kt
@@ -1,0 +1,13 @@
+package woowacourse.movie.view
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+abstract class BaseActivity<P> : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        initializePresenter()
+    }
+
+    abstract fun initializePresenter(): P
+}
