+++ app/src/main/java/woowacourse/movie/view/BaseView.kt
@@ -1,0 +1,5 @@
+package woowacourse.movie.view
+
+interface BaseView {
+    fun handleError()
+}
