+++ app/src/main/java/woowacourse/movie/view/MainActivity.kt
@@ -1,0 +1,26 @@
+package woowacourse.movie.view
+
+import android.os.Bundle
+import android.widget.ListView
+import woowacourse.movie.R
+import woowacourse.movie.conrtract.MainContract
+import woowacourse.movie.presenter.MainPresenter
+
+class MainActivity : BaseActivity<MainContract.Presenter>(), MainContract.View {
+    private val mainPresenter: MainContract.Presenter by lazy { initializePresenter() }
+    private val movieContentList: ListView by lazy { findViewById(R.id.movie_content_list) }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        mainPresenter.saveMovieContent()
+        setUpMovieContentListAdapter()
+    }
+
+    override fun initializePresenter() = MainPresenter()
+
+    override fun setUpMovieContentListAdapter() {
+        movieContentList.adapter = MovieContentListAdapter(this)
+    }
+}
