+++ app/src/main/res/layout/activity_movie_reservation_complete.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".MovieReservationCompleteActivity">
+    tools:context=".view.MovieReservationCompleteActivity">
 
     <TextView
         android:layout_width="match_parent"
